Refactor tests moving them to more specific paths
Split tests into proper files
